Refactor georoutingv2.js and utils.js for improved cookie handling and akamai code retrieval
- Updated georoutingv2.js to replace getAkamaiCode function with a new implementation that fetches akamai locale more efficiently.
- Moved getCookie function to utils.js for better accessibility and reusability.
- Enhanced error handling in the akamai code retrieval process to ensure robustness.
