+++ libs/features/georoutingv2/georoutingv2.js
@@ -1,4 +1,4 @@
-import { getFederatedContentRoot, getAkamaiCode } from '../../utils/utils.js';
+import { getFederatedContentRoot, getCookie } from '../../utils/utils.js';
 
 const OLD_GEOROUTING = 'oldgeorouting';
 
@@ -70,10 +70,30 @@ const [handleOverflow, removeOverflow] = (() => {
   ];
 })();
 
-export const getCookie = (name) => document.cookie
-  .split('; ')
-  .find((row) => row.startsWith(`${name}=`))
-  ?.split('=')[1];
+export const getAkamaiCode = () => new Promise((resolve, reject) => {
+  let akamaiLocale = null;
+  const urlParams = new URLSearchParams(window.location.search);
+  akamaiLocale = urlParams.get('akamaiLocale') || sessionStorage.getItem('akamai');
+
+  if (akamaiLocale !== null) {
+    resolve(akamaiLocale.toLowerCase());
+  } else {
+    /* c8 ignore next 5 */
+    fetch('https://geo2.adobe.com/json/', { cache: 'no-cache' }).then((resp) => {
+      if (resp.ok) {
+        resp.json().then((data) => {
+          const code = data.country.toLowerCase();
+          sessionStorage.setItem('akamai', code);
+          resolve(code);
+        });
+      } else {
+        reject(new Error(`Something went wrong getting the akamai Code. Response status text: ${resp.statusText}`));
+      }
+    }).catch((error) => {
+      reject(new Error(`Something went wrong getting the akamai Code. ${error.message}`));
+    });
+  }
+});
 
 // Determine if any of the locales can be linked to.
 async function getAvailableLocales(locales) {
